+++ packages/paraglide/paraglide-next/examples/app/src/lib/i18n.ts
@@ -1,16 +1,8 @@
-import { createI18n } from "@inlang/paraglide-next"
+import { Navigation, Middleware, PrefixStrategy } from "@inlang/paraglide-next"
 import type { AvailableLanguageTag } from "@/paraglide/runtime"
 import * as m from "@/paraglide/messages"
 
-export const {
-	Link,
-	middleware,
-	useRouter,
-	usePathname,
-	redirect,
-	permanentRedirect,
-	localizePath,
-} = createI18n<AvailableLanguageTag>({
+export const strategy = PrefixStrategy<AvailableLanguageTag>({
 	pathnames: {
 		"/about": m.about_path,
 		"/admin/[...rest]": {
@@ -19,5 +11,11 @@ export const {
 			"de-CH": "/administrator/[...rest]",
 		},
 	},
-	exclude: ["/not-translated"], //makes sure that the /not-translated page is not translated
+	prefixDefault: "never",
+	exclude: () => false,
+})
+
+export const middleware = Middleware({ strategy })
+export const { Link, useRouter, usePathname, redirect, permanentRedirect } = Navigation({
+	strategy,
 })