+++ packages/paraglide/paraglide-next/src/app/navigation/navigation.client.ts
@@ -1,40 +1,37 @@
 import * as NextNavigation from "next/navigation"
-import { setLanguageTag } from "$paraglide/runtime.js"
-import { addBasePath, basePath } from "./utils/basePath"
-import { RoutingStragey } from "./routing/interface"
-import { createLocaliseHref } from "./localiseHref"
-import { serializeCookie } from "./utils/cookie"
-import { LANG_COOKIE } from "./constants"
+import { languageTag, setLanguageTag } from "$paraglide/runtime.js"
+import { addBasePath, basePath } from "../utils/basePath"
+import type { RoutingStrategy } from "../routing-strategy/interface"
+import { createLocaliseHref } from "../localiseHref"
+import { serializeCookie } from "../utils/cookie"
+import { LANG_COOKIE } from "../constants"
+import { createRedirects } from "./redirect"
+import { createLink } from "./Link"
 
-export type LocalisedNavigation<T extends string> = ReturnType<typeof createNavigation<T>>
+export type LocalisedNavigation<T extends string> = ReturnType<typeof Navigation<T>>
 
-export const createNavigation = <T extends string>(
-	languageTag: () => T,
-	strategy: RoutingStragey<T>
-) => {
+export const Navigation = <T extends string>({ strategy }: { strategy: RoutingStrategy<T> }) => {
 	const localiseHref = createLocaliseHref(strategy)
-
-	type NextUsePathname = (typeof NextNavigation)["usePathname"]
 
 	/**
 	 * Get the current **non-localised** pathname. For example usePathname() on /de/dashboard?foo=bar would return "/dashboard"
 	 */
-	const usePathname: NextUsePathname = (...args) => {
-		const encodedLocalisedPathname = NextNavigation.usePathname(...args)
-		const localisedPathname = decodeURI(encodedLocalisedPathname)
-		return strategy.getCanonicalPath(localisedPathname, languageTag())
+	const usePathname = (): `/${string}` => {
+		const encodedLocalisedPathname = NextNavigation.usePathname()
+		const localisedPathname = decodeURI(encodedLocalisedPathname) as `/${string}`
+		return strategy.getCanonicalPath(localisedPathname, languageTag() as T)
 	}
 
-	/**
+	/**s
 	 * Get the router methods. For example router.push('/dashboard')
 	 */
 	const useRouter = () => {
 		const nextRouter = NextNavigation.useRouter()
-		const localisedCurrentPathname = usePathname()
+		const localisedCurrentPathname = NextNavigation.usePathname() as `/${string}`
 		const searchParams = NextNavigation.useSearchParams()
 		const canonicalCurrentPathname = strategy.getCanonicalPath(
 			localisedCurrentPathname,
-			languageTag()
+			languageTag() as T
 		)
 
 		type NavigateOptions = Parameters<(typeof nextRouter)["push"]>[1]
@@ -46,16 +43,23 @@ export const createNavigation = <T extends string>(
 		 * Navigate to the provided href. Pushes a new history entry.
 		 */
 		const push = (
-			canonicalPath: string,
+			canonicalDestinationPath: `/${string}`,
 			options?: (NavigateOptions & OptionalLanguageOption) | undefined
 		) => {
-			const locale = options?.locale ?? languageTag()
-			const localisedPath = localiseHref(canonicalPath, locale)
+			const locale = options?.locale ?? (languageTag() as T)
+			const isLanguageSwitch = locale !== languageTag()
+
+			const localisedPath = localiseHref(
+				canonicalDestinationPath,
+				locale,
+				canonicalCurrentPathname,
+				isLanguageSwitch
+			)
 
 			// If the current and new canonical paths are the same, but the language is different,
 			// we need to do a native reload to make sure the new language is used
 			if (
-				canonicalCurrentPathname === canonicalPath &&
+				canonicalCurrentPathname === canonicalDestinationPath &&
 				options?.locale &&
 				options.locale !== languageTag()
 			) {
@@ -69,7 +73,7 @@ export const createNavigation = <T extends string>(
 				document.cookie = serializeCookie({
 					...LANG_COOKIE,
 					value: locale,
-					path: basePath,
+					Path: basePath,
 				})
 
 				window.location.reload()
@@ -88,16 +92,22 @@ export const createNavigation = <T extends string>(
 		 * Navigate to the provided href. Replaces the current history entry.
 		 */
 		const replace = (
-			canonicalPath: string,
+			canonicalDestinationPath: `/${string}`,
 			options?: (NavigateOptions & OptionalLanguageOption) | undefined
 		) => {
-			const locale = options?.locale ?? languageTag()
-			const localisedPath = localiseHref(canonicalPath, locale)
+			const locale = options?.locale ?? (languageTag() as T)
+			const isLanguageSwitch = locale !== languageTag()
+			const localisedPath = localiseHref(
+				canonicalDestinationPath,
+				locale,
+				canonicalCurrentPathname,
+				isLanguageSwitch
+			)
 
 			// If the current and new canonical paths are the same, but the language is different,
 			// we need to do a native reload to make sure the new language is used
 			if (
-				canonicalCurrentPathname === canonicalPath &&
+				canonicalCurrentPathname === canonicalDestinationPath &&
 				options?.locale &&
 				options.locale !== languageTag()
 			) {
@@ -111,7 +121,7 @@ export const createNavigation = <T extends string>(
 				document.cookie = serializeCookie({
 					...LANG_COOKIE,
 					value: locale,
-					path: basePath,
+					Path: basePath,
 				})
 
 				window.location.reload()
@@ -129,9 +139,13 @@ export const createNavigation = <T extends string>(
 		/**
 		 * Prefetch the provided href.
 		 */
-		const prefetch = (canonicalPath: string, options: PrefetchOptions & OptionalLanguageOption) => {
-			const locale = options?.locale ?? languageTag()
-			const localisedPath = localiseHref(canonicalPath, locale)
+		const prefetch = (
+			canonicalDestinationPath: `/${string}`,
+			options: PrefetchOptions & OptionalLanguageOption
+		) => {
+			const locale = options?.locale ?? (languageTag() as T)
+			const isLanguageSwitch = locale !== languageTag()
+			const localisedPath = localiseHref(canonicalDestinationPath, locale, "/", isLanguageSwitch)
 			return nextRouter.prefetch(localisedPath, options)
 		}
 
@@ -143,5 +157,10 @@ export const createNavigation = <T extends string>(
 		}
 	}
 
-	return { useRouter, usePathname }
+	return {
+		useRouter,
+		usePathname,
+		...createRedirects(languageTag, strategy),
+		Link: createLink(strategy),
+	}
 }