+++ packages/paraglide/paraglide-next/src/app/navigation/navigation.server.ts
@@ -1,10 +1,18 @@
 import type { LocalisedNavigation } from "./navigation.client"
+import type { RoutingStrategy } from "../routing-strategy/interface"
+import { createRedirects } from "./redirect"
+import { languageTag } from "$paraglide/runtime.js"
+import { createLink } from "./Link"
 
 /**
  * Implements the same API as NextNavigation, but throws an error when used.
  * Usefull for poisoning the client-side navigation hooks on the server.
  */
-export function createNoopNavigation<T extends string>(): LocalisedNavigation<T> {
+export function Navigation<T extends string>({
+	strategy,
+}: {
+	strategy: RoutingStrategy<T>
+}): LocalisedNavigation<T> {
 	return {
 		usePathname: () => {
 			throw new Error("usePathname is not available on the server")
@@ -12,5 +20,7 @@ export function createNoopNavigation<T extends string>(): LocalisedNavigation<T>
 		useRouter: () => {
 			throw new Error("useRouter is not available on the server")
 		},
+		...createRedirects(languageTag, strategy),
+		Link: createLink(strategy),
 	}
 }